refactor: simplify QuickExpenseModal by extracting header and logic into separate components
